# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountResponse": {
        "description": "One account as returned to the client.\n\nA deliberately narrow projection of :class:`~traccio.domain.models.Account`:\n``user_id`` (implied by the caller) and ``identification_hash`` (an internal\nmatching detail) are intentionally omitted.\n\nAttributes\n----------\nid : UUID\n    Stable account identifier.\nconnection_id : UUID or None\n    Connection currently exposing this account, or ``None`` for a manual\n    account (ADR 0020).\nsource : AccountSource\n    ``synced`` if the account is backed by a bank connection, ``manual`` if\n    the user created and maintains it. Derived from ``connection_id`` (see\n    :func:`~traccio.domain.accounts.account_source`), never stored \u2014 sent so\n    the client does not infer it from a null.\nkind : AccountKind\n    ``current``, ``savings``, ``card``, ``wallet``, ``cash``, or\n    ``voucher``.\ncurrency : str\n    The account's ISO 4217 currency.\nname : str or None\n    Provider-supplied display name (overwritten on every sync).\nalias : str or None\n    User-chosen display name (ADR 0017), or ``None`` if unset.\ndisplay_name : str or None\n    The one name the client should actually show \u2014 ``alias`` if set, else\n    ``name``, else ``None`` \u2014 resolved once by\n    :func:`~traccio.domain.accounts.display_name` so the client does not\n    reimplement the fallback (it previously did, inconsistently, in three\n    different places).\ncolor : PaletteColor or None\n    User-chosen colour, or ``None`` if unset.\nicon : AccountIcon or None\n    User-chosen icon, or ``None`` if unset.\ncreated_at : datetime\n    When the account was first recorded.",
        "properties": {
          "alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alias"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaletteColor"
              },
              {
                "type": "null"
              }
            ]
          },
          "connection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "icon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountIcon"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AccountKind"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source": {
            "$ref": "#/components/schemas/AccountSource"
          }
        },
        "required": [
          "id",
          "connection_id",
          "source",
          "kind",
          "currency",
          "name",
          "alias",
          "display_name",
          "color",
          "icon",
          "created_at"
        ],
        "title": "AccountResponse",
        "type": "object"
      },
...
      "AccountsResponse": {
        "description": "Envelope for the account list.\n\nA wrapper object rather than a bare array leaves room for pagination or\nmetadata later without breaking the generated Swift client.\n\nAttributes\n----------\naccounts : list[AccountResponse]\n    The caller's accounts, oldest first.",
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountResponse"
            },
            "title": "Accounts",
            "type": "array"
          }
        },
        "required": [
          "accounts"
        ],
        "title": "AccountsResponse",
        "type": "object"
      },
...
      "CreateManualAccountRequest": {
        "description": "Body for creating a manual account (ADR 0020).\n\nA manual account has no bank connection: the user names it (``alias``),\npicks what kind of account it is and its currency, and optionally an\nappearance. ``name``/``identification_hash``/``connection_id`` are all\nserver-side ``None`` for it.\n\nAttributes\n----------\nalias : str\n    The account's name, e.g. ``\"Contanti\"``. Stripped and validated by\n    :func:`~traccio.domain.accounts.normalize_account_alias`; a manual\n    account must have one (unlike a synced account, which falls back to the\n    provider ``name``).\nkind : AccountKind\n    What type of account it is \u2014 typically ``cash`` or ``wallet`` for a\n    manual one, but any kind is accepted.\ncurrency : str\n    The account's ISO 4217 currency (three uppercase letters).\ncolor : PaletteColor or None\n    Optional colour, from the fixed :class:`PaletteColor` vocabulary.\nicon : AccountIcon or None\n    Optional icon, from the fixed :class:`AccountIcon` vocabulary.",
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaletteColor"
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "icon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountIcon"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/AccountKind"
          }
        },
        "required": [
          "alias",
          "kind",
          "currency"
        ],
        "title": "CreateManualAccountRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaletteColor": {
        "description": "A semantic colour, shared by :class:`~traccio.domain.models.Account` and\n:class:`~traccio.domain.models.Category`.\n\nDeliberately a fixed vocabulary rather than a free hex string (ADR 0017):\nthe client's ``Colors.xcassets`` is the only place a colour has an explicit\ndark-mode variant, so a user-chosen hex would have no dark counterpart and\nsilently break the design system's dark-mode contract. One shared enum for\nboth entities, not two, because there is exactly one colour vocabulary in\nthe app \u2014 the icon vocabularies differ (see :class:`AccountIcon` and\n``CategoryIcon``), the colour one does not.\n\nAttributes\n----------\nBLUE, INDIGO, PURPLE, PINK, RED, ORANGE, AMBER, GREEN, TEAL, SLATE : str\n    The ten selectable tones. ``SLATE`` is the neutral default for\n    anything the user has not deliberately coloured yet.",
        "enum": [
          "blue",
          "indigo",
          "purple",
          "pink",
          "red",
          "orange",
          "amber",
          "green",
          "teal",
          "slate"
        ],
        "title": "PaletteColor",
        "type": "string"
      },
...
      "RenameAccountRequest": {
        "description": "Body for setting or clearing an account's alias.\n\nAttributes\n----------\nalias : str or None\n    The new alias, stripped and validated by\n    :func:`~traccio.domain.accounts.normalize_account_alias`. ``None``\n    clears the alias and falls back to the provider name \u2014 a mandatory,\n    nullable field rather than an optional one, so \"absent\" and \"clear\n    it\" are never ambiguous.",
        "properties": {
          "alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alias"
          }
        },
        "required": [
          "alias"
        ],
        "title": "RenameAccountRequest",
        "type": "object"
      },
...
      "SetAccountAppearanceRequest": {
        "description": "Body for setting an account's colour and icon.\n\nA full replace, not a partial update \u2014 both fields are mandatory (but\nindividually nullable, to allow clearing) \u2014 mirroring\n``appearance``'s single action-style endpoint rather than two.\n\nAttributes\n----------\ncolor : PaletteColor or None\n    The new colour, or ``None`` to clear it.\nicon : AccountIcon or None\n    The new icon, or ``None`` to clear it.",
        "properties": {
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaletteColor"
              },
              {
                "type": "null"
              }
            ]
          },
          "icon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountIcon"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "color",
          "icon"
        ],
        "title": "SetAccountAppearanceRequest",
        "type": "object"
      },
      "SetAccountKindRequest": {
        "description": "Body for reclassifying a manual account's ``kind``.\n\nThe only way to turn an existing manual account into a meal-voucher\naccount (ADR 0029) \u2014 e.g. a \"Buoni Pasto\" account created ``cash`` before\nthe feature existed, or by the Satispay import (ADR 0023). Manual only;\na synced account's ``kind`` is provider-derived (``409\naccount_not_manual``).\n\nAttributes\n----------\nkind : AccountKind\n    The new kind.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AccountKind"
          }
        },
        "required": [
          "kind"
        ],
        "title": "SetAccountKindRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Traccio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/accounts": {
      "get": {
        "description": "List the current user's accounts.\n\nScoped to the current user (see :func:`~traccio.api.deps.current_user_id`).\nThe query itself is already written ``scoped by user_id``; only the source\nof the id changes when auth arrives.\n\nParameters\n----------\nsession : Session\n    Request-scoped database session (see :func:`get_session`).\nuser_id : UUID\n    The user whose accounts to return.\n\nReturns\n-------\nAccountsResponse\n    The user's accounts, oldest first.",
        "operationId": "accounts_accounts_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accounts"
      },
      "post": {
        "description": "Create a manual account \u2014 one with no bank connection (ADR 0020).\n\nFor hand-tracked money: a cash float, an investment pass-through. The\naccount has ``connection_id`` and ``identification_hash`` both ``None``, so\na sync can never touch it, and its ``source`` is ``manual``. A ``422`` if\nthe alias is blank or too long, or the currency/kind is not a recognized\nvalue.\n\nParameters\n----------\nbody : CreateManualAccountRequest\n    The new account's name, kind, currency, and optional appearance.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the account belongs to.\n\nReturns\n-------\nAccountResponse\n    The newly created manual account.",
        "operationId": "create_manual_account_endpoint_accounts_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Manual Account Endpoint"
      }
    },
    "/accounts/{account_id}": {
      "delete": {
        "description": "Delete a manual account (ADR 0020).\n\nA ``404`` if the account is unknown or not the caller's. A ``409\naccount_not_manual`` if it is a synced account \u2014 those are removed only by\nthe connection flow, never here. A ``409 account_not_empty`` if it still\nholds any transaction: the account's movements must be deleted first\n(``DELETE /transactions/{id}``), mirroring ``409 category_in_use`` \u2014\nfinancial data is never cascade-deleted out from under the user.\n\nParameters\n----------\naccount_id : UUID\n    The account to delete.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the account belongs to.",
        "operationId": "delete_manual_account_endpoint_accounts__account_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Manual Account Endpoint"
      }
    },
    "/accounts/{account_id}/appearance": {
      "post": {
        "description": "Set an account's colour and icon.\n\nA full replace: both fields are applied together. A ``404`` if the\naccount is unknown or not the caller's; an unrecognized value is\nrejected by request validation before the handler runs (``422``).\n\nParameters\n----------\naccount_id : UUID\n    The account to restyle.\nbody : SetAccountAppearanceRequest\n    The new colour and icon.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the account belongs to.\n\nReturns\n-------\nAccountResponse\n    The account under its new appearance.",
        "operationId": "set_account_appearance_endpoint_accounts__account_id__appearance_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAccountAppearanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Account Appearance Endpoint"
      }
    },
    "/accounts/{account_id}/kind": {
      "post": {
        "description": "Reclassify a manual account's ``kind`` (ADR 0029).\n\nThe only way to turn an existing account into a meal-voucher one \u2014 e.g.\na \"Buoni Pasto\" account created ``cash`` before the feature existed, or\nby the Satispay import (ADR 0023) \u2014 without deleting and recreating it\n(which ``DELETE /accounts/{id}`` refuses once it holds movements, ``409\naccount_not_empty``). A ``404`` if the account is unknown or not the\ncaller's. A ``409 account_not_manual`` if it is a synced account \u2014 a\nsynced account's ``kind`` is provider-derived and only a sync may write\nit (``upsert_account``).\n\nParameters\n----------\naccount_id : UUID\n    The account to reclassify.\nbody : SetAccountKindRequest\n    The new kind.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the account belongs to.\n\nReturns\n-------\nAccountResponse\n    The account under its new kind.",
        "operationId": "set_account_kind_endpoint_accounts__account_id__kind_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAccountKindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Account Kind Endpoint"
      }
    },
    "/accounts/{account_id}/rename": {
      "post": {
        "description": "Set or clear an account's alias.\n\nA ``404`` if the account is unknown or not the caller's; a ``422`` if the\nalias is blank (use ``null`` to clear it) or too long.\n\nParameters\n----------\naccount_id : UUID\n    The account to rename.\nbody : RenameAccountRequest\n    The new alias, or ``null`` to clear it.\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the account belongs to.\n\nReturns\n-------\nAccountResponse\n    The account under its new alias.",
        "operationId": "rename_account_endpoint_accounts__account_id__rename_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Account Endpoint"
      }
    },
...
  }
}